If the Car dont have a Trunk
If you use a Vehicle like as example the Burrito, the Burrito has no Trunk, so it doesnt show up the 3dText to take the toolkit. Thats why you have to do some changes, that im posted at the Readme. So its simple. If you have any Questions, feel free to ask.